pyproject.toml
[project]
name = "Nero"
version = "0.1.0"
description = "Options Analysis Tool with OCR capabilities"
authors = [
    {name = "Dylan Lam", email = "dylanclam12@gmail.com"},
]
dependencies = [
    "streamlit>=1.28.0",
    "ipykernel>=6.29.5", 
    "pandas>=2.3.0", 
    "opencv-python>=4.11.0.86", 
    "matplotlib>=3.10.3", 
    "pytesseract>=0.3.13", 
    "easyocr>=1.7.2", 
    "img2table>=1.4.1", 
    "numpy==2.2",
    "pillow>=10.0.0"
]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
